Add files via upload
This is the original game that I have decided to enhance. It was created in 2022 on PyCharm using the Python language. It was created as a text-based educational game to help kids learn how to read, practice with directions, and problem solve. It showcases my coding abilities at a basic level, and is perfect to enhance in the 3 categories: software engineering, data structures and algorithms, and databases.