Add files via upload
In Module 1 I stated that I would enhance this game to show my understanding of 
software engineering, data structures and algorithms, and databases. I have exceeded my 
expectations by working very hard each week to make sure to do showcase my skills. For the 
category of databases this week I added a to-do list where I can utilize CRUD and make sure that 
it flowed well with the game. I feel like the enhancements that I made align perfectly with the 
goals that I had made.
When I modified the Shape Monster Game to have database functionalities, I learned 
how important structuring data is. The biggest challenge that I faced was making sure that the todo list was user-friendly and allowed the users to interact with the database. I tested over and over with my son and realized that I needed to create a menu in order to use the to-do list to 
during the game play. It worked great after several tests. This process helped me improve my 
problem-solving skills and provided knowledge on integrating database functionalities in a 
software project