Add files via upload
did meet the course outcomes that I planned to meet with this enhancement in Module One by 
successfully converting the game to C++. The extra enhancements that I did improve some of the 
information for Algorithms and Data Structures, as well as the game play experience. I would 
say that I exceeded my expectations.

I learned the importance of feedback. I tested the game and debugged several times before it was 
without errors in C++. Once I had it without errors, I tested the game with my son. At that point I 
had to redo a lot because it wasn’t working or looking how I wanted it to. After I got it formatted 
correctly, I asked him what he would do to improve the game. He gave me a great suggestion. The suggestion was to make the monster randomly appear in different rooms. I immediately took his suggestion and updated the game for him to play.